Move flocker.node._model and flocker.node._config to flocker.control where they belong now.
