+++ flocker/node/__init__.py
@@ -4,30 +4,8 @@
 Local node manager for Flocker.
 """
 
-from ._config import (
-    FlockerConfiguration, ConfigurationError, FigConfiguration,
-    applications_to_flocker_yaml, model_from_configuration,
-    current_from_configuration,
-    )
-from ._model import (
-    Application, Deployment, DockerImage, Node, Port, Link, AttachedVolume,
-    NodeState)
 from ._deploy import Deployer
 
 __all__ = [
-    'FlockerConfiguration',
-    'ConfigurationError',
-    'applications_to_flocker_yaml',
-    'current_from_configuration',
-    'model_from_configuration',
-    'Application',
-    'Deployment',
     'Deployer',
-    'DockerImage',
-    'FigConfiguration',
-    'Node',
-    'Port',
-    'Link',
-    'AttachedVolume',
-    'NodeState',
 ]
