+++ flocker/node/functional/test_docker.py
@@ -27,7 +27,7 @@ from ..test.test_docker import make_idockerclient_tests
 from .._docker import (
     DockerClient, PortMap, Environment, NamespacedDockerClient,
     BASE_NAMESPACE, Volume)
-from .._model import RestartNever, RestartAlways, RestartOnFailure
+from ...control._model import RestartNever, RestartAlways, RestartOnFailure
 from ..testtools import if_docker_configured, wait_for_unit_state
